+++ packages/core/src/connectivity/scp-cf/destination/destination-from-service.ts
@@ -1,10 +1,6 @@
-import { createLogger, encodeBase64 } from '@sap-cloud-sdk/util';
+import { createLogger } from '@sap-cloud-sdk/util';
+import { JwtPayload } from 'jsonwebtoken';
 import { decodeJwt, isUserToken, JwtPair, verifyJwt } from '../jwt';
-import {
-  addProxyConfigurationInternet,
-  ProxyStrategy,
-  proxyStrategy
-} from '../../../http-agent/proxy-util';
 import { IsolationStrategy } from '../cache';
 import { jwtBearerToken, serviceToken } from '../token-accessor';
 import { addProxyConfigurationOnPrem } from '../connectivity-service';
@@ -18,6 +14,12 @@ import {
   DestinationServiceCredentials,
   XsuaaServiceCredentials
 } from '../environment-accessor-types';
+import { getSubdomainAndZoneId } from '../xsuaa-service';
+import {
+  addProxyConfigurationInternet,
+  ProxyStrategy,
+  proxyStrategy
+} from '../../../http-agent';
 import { Destination } from './destination-service-types';
 import {
   alwaysProvider,
@@ -40,7 +42,7 @@ type RequiredProperties<T, P extends keyof T> = Required<Pick<T, P>> &
   Omit<T, P>;
 
 const logger = createLogger({
-  package: 'core',
+  package: 'connectivity',
   messageContext: 'destination-accessor-service'
 });
 
@@ -48,6 +50,11 @@ interface DestinationSearchResult {
   destination: Destination;
   fromCache: boolean;
   origin: DestinationOrigin;
+}
+
+interface SubscriberTokens {
+  userJwt?: JwtPair; // is undefined when iss is passed ad token
+  serviceJwt: JwtPair;
 }
 
 const emptyDestinationByType: DestinationsByType = {
@@ -77,6 +84,9 @@ export async function getDestinationFromDestinationService(
   );
 }
 
+/**
+ * @internal
+ */
 class DestinationFromServiceRetriever {
   public static async getDestinationFromDestinationService(
     name: string,
@@ -87,10 +97,11 @@ class DestinationFromServiceRetriever {
 
     const xsuaaCredentials = getXsuaaServiceCredentials(options.userJwt);
     const providerToken =
-      await DestinationFromServiceRetriever.getProviderClientCredentialsToken(
+      await DestinationFromServiceRetriever.getProviderServiceToken(
         xsuaaCredentials,
         options
       );
+
     const da = new DestinationFromServiceRetriever(
       name,
       { ...options, xsuaaCredentials },
@@ -110,31 +121,26 @@ class DestinationFromServiceRetriever {
 
     let { destination } = destinationResult;
 
-    if (destination.authentication === 'OAuth2SAMLBearerAssertion') {
-      /*
-    This covers the two technical user propagation cases https://help.sap.com/viewer/cca91383641e40ffbe03bdc78f00f681/Cloud/en-US/3cb7b81115c44cf594e0e3631291af94.html (fetchDestinationBySystemUser)
-    If no system user is set the subscriber or provider destination is fetched depending on the content of the JWT (fetchDestinationByUserJwt)
-    */
-      destination =
-        (await da.fetchDestinationBySystemUser(destinationResult)) ||
-        (await da.fetchDestinationByUserJwt());
-    }
-    if (destination.authentication === 'OAuth2UserTokenExchange') {
-      destination = await da.fetchDestinationByUserTokenExchange(
-        destinationResult.origin
+    if (
+      destination.authentication === 'OAuth2UserTokenExchange' ||
+      destination.authentication === 'OAuth2JWTBearer' ||
+      (destination.authentication === 'OAuth2SAMLBearerAssertion' &&
+        !da.usesSystemUser(destination))
+    ) {
+      destination = await da.fetchDestinationWithUserExchangeFlows(
+        destinationResult
       );
     }
+
     if (
+      destination.authentication === 'OAuth2Password' ||
+      destination.authentication === 'ClientCertificateAuthentication' ||
       destination.authentication === 'OAuth2ClientCredentials' ||
-      destination.authentication === 'OAuth2Password'
+      da.usesSystemUser(destination)
     ) {
-      destination = await da.fetchDestinationByClientCrendentialsGrant();
-    }
-    if (destination.authentication === 'OAuth2JWTBearer') {
-      destination = await da.fetchDestinationByUserJwt();
-    }
-    if (destination.authentication === 'ClientCertificateAuthentication') {
-      destination = await da.addClientCertAuth(destinationResult.origin);
+      destination = await da.fetchDestinationWithNonUserExchangeFlows(
+        destinationResult
+      );
     }
 
     const withProxySetting = await da.addProxyConfiguration(destination);
@@ -144,16 +150,22 @@ class DestinationFromServiceRetriever {
 
   private static async getSubscriberToken(
     options: DestinationOptions
-  ): Promise<JwtPair | undefined> {
+  ): Promise<SubscriberTokens | undefined> {
     if (options.userJwt) {
       if (options.iss) {
         logger.warn(
           'You have provided the `userJwt` and `iss` options to fetch the destination. This is most likely unintentional. Ignoring `iss`.'
         );
       }
+      const encoded = await serviceToken('destination', {
+        ...options
+      });
       return {
-        decoded: await verifyJwt(options.userJwt, options),
-        encoded: options.userJwt
+        userJwt: {
+          decoded: await verifyJwt(options.userJwt, options),
+          encoded: options.userJwt
+        },
+        serviceJwt: { encoded, decoded: decodeJwt(encoded) }
       };
     }
 
@@ -162,14 +174,16 @@ class DestinationFromServiceRetriever {
         'Using `iss` option to fetch a destination instead of a full JWT. No validation is performed.'
       );
       const payload = { iss: options.iss };
-      return {
-        decoded: payload,
-        encoded: encodeBase64(JSON.stringify(payload))
-      };
+      const encoded = await serviceToken('destination', {
+        ...options,
+        userJwt: payload
+      });
+      const clientCertJwt = { encoded, decoded: decodeJwt(encoded) };
+      return { serviceJwt: clientCertJwt };
     }
   }
 
-  private static async getProviderClientCredentialsToken(
+  private static async getProviderServiceToken(
     xsuaaCredentials: XsuaaServiceCredentials,
     options: DestinationOptions
   ): Promise<JwtPair> {
@@ -182,17 +196,6 @@ class DestinationFromServiceRetriever {
     return { encoded, decoded: decodeJwt(encoded) };
   }
 
-  private static async getSubscriberClientCredentialsToken(
-    options: DestinationOptions
-  ): Promise<string> {
-    if (!options.userJwt) {
-      throw new Error(
-        'User JWT is needed to obtain a client credentials token for the subscriber account.'
-      );
-    }
-    return serviceToken('destination', options);
-  }
-
   private options: RequiredProperties<
     Omit<DestinationOptions, 'userJwt' | 'iss'>,
     'isolationStrategy' | 'selectionStrategy' | 'useCache'
@@ -201,8 +204,8 @@ class DestinationFromServiceRetriever {
   private constructor(
     readonly name: string,
     options: DestinationOptions & { xsuaaCredentials: XsuaaServiceCredentials },
-    readonly subscriberToken: JwtPair | undefined,
-    readonly providerClientCredentialsToken: JwtPair
+    readonly subscriberToken: SubscriberTokens | undefined,
+    readonly providerServiceToken: JwtPair
   ) {
     const defaultOptions = {
       isolationStrategy: IsolationStrategy.Tenant_User,
@@ -292,138 +295,124 @@ class DestinationFromServiceRetriever {
     );
   }
 
-  private async getAuthTokenForOAuth2UserTokenExchange(
-    destinationOrigin: DestinationOrigin
+  private getExchangeTenant(destination: Destination): string | undefined {
+    if (destination.authentication !== 'OAuth2ClientCredentials') {
+      return undefined;
+    }
+    if (destination.originalProperties!['tokenServiceURLType'] !== 'Common') {
+      return undefined;
+    }
+    const subdomainSubscriber = getSubdomainAndZoneId(
+      this.subscriberToken?.userJwt?.encoded
+    ).subdomain;
+    const subdomainProvider = getSubdomainAndZoneId(
+      this.providerServiceToken?.encoded
+    ).subdomain;
+    return subdomainSubscriber || subdomainProvider || undefined;
+  }
+
+  private async getAuthTokenForOAuth2ClientCrendentials(
+    destinationResult: DestinationSearchResult
   ): Promise<AuthAndExchangeTokens> {
-    if (!isUserToken(this.subscriberToken)) {
-      throw Error(
-        'No user token (JWT) has been provided. This is strictly necessary for `OAuth2UserTokenExchange`.'
+    const { destination, origin } = destinationResult;
+    // This covers the X-Tenant case https://api.sap.com/api/SAP_CP_CF_Connectivity_Destination/resource
+    const exchangeTenant = this.getExchangeTenant(destination);
+    const clientGrant = await serviceToken('destination', {
+      userJwt:
+        origin === 'subscriber'
+          ? this.subscriberToken!.serviceJwt.decoded
+          : this.providerServiceToken.decoded
+    });
+    return { authHeaderJwt: clientGrant, exchangeTenant };
+  }
+
+  // This covers the two technical user propagation https://help.sap.com/viewer/cca91383641e40ffbe03bdc78f00f681/Cloud/en-US/3cb7b81115c44cf594e0e3631291af94.html
+  private usesSystemUser(destination: Destination): boolean {
+    // put this in the non user dependent block
+    if (
+      destination.systemUser &&
+      destination.authentication === 'OAuth2SAMLBearerAssertion'
+    ) {
+      logger.debug(
+        `System user found on destination: "${destination.name}". 
+The property SystemUser has been deprecated. 
+It is highly recommended that you stop using it.
+Possible alternatives for such technical user authentication are BasicAuthentication, OAuth2ClientCredentials, or ClientCertificateAuthentication`
       );
+      return true;
     }
 
-    // Case 1 Destination in provider and jwt issued for provider account
+    return false;
+  }
+
+  private async getAuthTokenForOAuth2UserBasedTokenExchanges(
+    destinationResult: DestinationSearchResult
+  ): Promise<AuthAndExchangeTokens> {
+    const { destination, origin } = destinationResult;
+    if (!this.subscriberToken || !isUserToken(this.subscriberToken.userJwt)) {
+      throw Error(
+        `No user token (JWT) has been provided. This is strictly necessary for '${destination.authentication}'.`
+      );
+    }
+    // This covers OAuth to user dependend auth flows https://help.sap.com/viewer/cca91383641e40ffbe03bdc78f00f681/Cloud/en-US/39d42654093e4f8db20398a06f7eab2b.html and https://api.sap.com/api/SAP_CP_CF_Connectivity_Destination/resource
+    // Which is the same for: OAuth2UserTokenExchange, OAuth2JWTBearer and OAuth2SAMLBearerAssertion
+
+    // Case 1 Destination in provider and jwt issued for provider account -> not extra x-user-token header needed
     if (this.isProviderAndSubscriberSameTenant()) {
       logger.debug(
-        `OAuth2UserTokenExchange flow started without user exchange token for destination ${this.name} of the provider account.`
+        `UserExchange flow started without user exchange token for destination ${this.name} of the provider account.`
       );
       return {
         authHeaderJwt: await jwtBearerToken(
-          this.subscriberToken.encoded,
+          this.subscriberToken.userJwt.encoded,
           getDestinationService(),
           this.options
         )
       };
     }
-    // Case 2 Destination in provider and jwt issued for subscriber account
-    if (destinationOrigin === 'provider') {
-      logger.debug(
-        `OAuth2UserTokenExchange flow started for destination ${this.name} of the provider account.`
-      );
-      return {
-        authHeaderJwt: this.providerClientCredentialsToken.encoded,
-        exchangeHeaderJwt: this.subscriberToken.encoded
-      };
-    }
+    // Case 2 Subscriber and provider account not the same -> x-user-token  header passed to determine user and tenant in token service URL and service token to get the destination
+    const serviceJwt =
+      origin === 'subscriber'
+        ? this.subscriberToken.serviceJwt
+        : this.providerServiceToken;
+    logger.debug(
+      `UserExchange flow started for destination ${this.name} of the ${origin} account.`
+    );
 
-    // Case 3 Destination in subscriber and jwt issued for subscriber account
-    if (destinationOrigin === 'subscriber') {
-      logger.debug(
-        `OAuth2UserTokenExchange flow started for destination ${this.name} of the subscriber account.`
-      );
-      return {
-        authHeaderJwt:
-          await DestinationFromServiceRetriever.getSubscriberClientCredentialsToken(
-            this.options
-          ),
-        exchangeHeaderJwt: this.subscriberToken.encoded
-      };
-    }
+    return {
+      authHeaderJwt: serviceJwt.encoded, // token to get destination from service
+      exchangeHeaderJwt: this.subscriberToken.userJwt.encoded // token considered for user and tenant
+    };
+
     throw new Error(
-      'Not possible to build tokens for OAuth2UserTokenExchange flow.'
+      `Not possible to build tokens for ${destination.authentication} flow for destination ${destination.name}.`
     );
-  }
-
-  private async fetchDestinationByUserTokenExchange(
-    destinationOrigin: DestinationOrigin
-  ): Promise<Destination> {
-    // This covers the three OAuth2UserTokenExchange cases https://help.sap.com/viewer/cca91383641e40ffbe03bdc78f00f681/Cloud/en-US/39d42654093e4f8db20398a06f7eab2b.html
-    const token = await this.getAuthTokenForOAuth2UserTokenExchange(
-      destinationOrigin
-    );
-    return this.fetchDestinationByToken(token);
   }
 
   /**
-   * @hidden
+   * @internal
    * This method calls the 'find destination by name' endpoint of the destination service using a client credentials grant.
    * For the find by name endpoint, the destination service will take care of OAuth flows and include the token in the destination.
+   * @param destinationResult - Result of the getDestinations call for which the exchange flow is triggred
+   * @returns Destination containing the auth token.
    */
-  private async fetchDestinationByClientCrendentialsGrant(): Promise<Destination> {
-    const clientGrant = await serviceToken('destination', {
-      userJwt:
-        this?.subscriberToken?.decoded ||
-        this.providerClientCredentialsToken.decoded
-    });
-
-    return this.fetchDestinationByToken(clientGrant);
-  }
-
-  private async fetchDestinationBySystemUser(
+  private async fetchDestinationWithNonUserExchangeFlows(
     destinationResult: DestinationSearchResult
-  ): Promise<Destination | undefined> {
-    if (destinationResult.destination.systemUser) {
-      const token =
-        destinationResult.origin === 'provider'
-          ? this.providerClientCredentialsToken.encoded
-          : await DestinationFromServiceRetriever.getSubscriberClientCredentialsToken(
-              this.options
-            );
-      logger.debug(
-        `System user found on destination: "${destinationResult.destination.name}".`
-      );
-
-      if (destinationResult.origin) {
-        return this.fetchDestinationByToken(token);
-      }
-    }
-  }
-
-  private async fetchDestinationByUserJwt(): Promise<Destination> {
-    const destinationService = getDestinationService();
-
-    /* This covers the two business user propagation cases https://help.sap.com/viewer/cca91383641e40ffbe03bdc78f00f681/Cloud/en-US/3cb7b81115c44cf594e0e3631291af94.html
-     The two cases are JWT issued from provider or JWT from subscriber - the two cases are handled automatically.
-     In the provider case the subdomain replacement in the xsuaa.url with the iss value does nothing but this does not hurt. */
-    if (!isUserToken(this.subscriberToken)) {
-      throw Error(
-        'No user token (JWT) has been provided. This is strictly necessary for principal propagation.'
-      );
-    }
-    const accessToken = await jwtBearerToken(
-      this.subscriberToken.encoded,
-      destinationService,
-      this.options
-    );
-    return this.fetchDestinationByToken(accessToken);
-  }
-
-  private async addClientCertAuth(
-    origin: DestinationOrigin
   ): Promise<Destination> {
-    const accessToken = await serviceToken('destination', {
-      ...this.options,
-      userJwt:
-        origin === 'subscriber'
-          ? this.subscriberToken!.decoded
-          : this.providerClientCredentialsToken.decoded
-    });
-
-    return fetchDestination(
-      this.destinationServiceCredentials.uri,
-      accessToken,
-      this.name,
-      this.options
+    const token = await this.getAuthTokenForOAuth2ClientCrendentials(
+      destinationResult
     );
+
+    return this.fetchDestinationByToken(token);
+  }
+
+  private async fetchDestinationWithUserExchangeFlows(
+    destinationResult: DestinationSearchResult
+  ): Promise<Destination> {
+    const token = await this.getAuthTokenForOAuth2UserBasedTokenExchanges(
+      destinationResult
+    );
+    return this.fetchDestinationByToken(token);
   }
 
   private async addProxyConfiguration(
@@ -431,7 +420,10 @@ class DestinationFromServiceRetriever {
   ): Promise<Destination> {
     switch (proxyStrategy(destination)) {
       case ProxyStrategy.ON_PREMISE_PROXY:
-        return addProxyConfigurationOnPrem(destination, this.subscriberToken);
+        return addProxyConfigurationOnPrem(
+          destination,
+          this.subscriberToken?.userJwt
+        );
       case ProxyStrategy.INTERNET_PROXY:
         return addProxyConfigurationInternet(destination);
       case ProxyStrategy.NO_PROXY:
@@ -443,6 +435,17 @@ class DestinationFromServiceRetriever {
     }
   }
 
+  // For iss token the userJwt may be undefined.
+  private selectSubscriberJwt(): JwtPayload {
+    if (!this.subscriberToken) {
+      throw new Error('Try to get subscriber token but value is undefined.');
+    }
+    return (
+      this.subscriberToken.userJwt?.decoded ||
+      this.subscriberToken.serviceJwt.decoded
+    );
+  }
+
   private updateDestinationCache(
     destination: Destination,
     destinationOrigin: DestinationOrigin
@@ -452,8 +455,8 @@ class DestinationFromServiceRetriever {
     }
     destinationCache.cacheRetrievedDestination(
       destinationOrigin === 'subscriber'
-        ? this.subscriberToken!.decoded
-        : this.providerClientCredentialsToken.decoded,
+        ? this.selectSubscriberJwt()
+        : this.providerServiceToken.decoded,
       destination,
       this.options.isolationStrategy
     );
@@ -463,7 +466,7 @@ class DestinationFromServiceRetriever {
     DestinationSearchResult | undefined
   > {
     const provider = await this.getInstanceAndSubaccountDestinations(
-      this.providerClientCredentialsToken.encoded
+      this.providerServiceToken.encoded
     );
     const destination = this.options.selectionStrategy(
       {
@@ -483,7 +486,7 @@ class DestinationFromServiceRetriever {
 
   private getProviderDestinationCache(): DestinationSearchResult | undefined {
     const destination = destinationCache.retrieveDestinationFromCache(
-      this.providerClientCredentialsToken.decoded,
+      this.providerServiceToken.decoded,
       this.name,
       this.options.isolationStrategy
     );
@@ -502,12 +505,8 @@ class DestinationFromServiceRetriever {
       );
     }
 
-    const accessToken = await serviceToken('destination', {
-      ...this.options,
-      userJwt: this.subscriberToken.decoded
-    });
     const subscriber = await this.getInstanceAndSubaccountDestinations(
-      accessToken
+      this.subscriberToken.serviceJwt.encoded
     );
     const destination = this.options.selectionStrategy(
       {
@@ -524,7 +523,7 @@ class DestinationFromServiceRetriever {
 
   private getSubscriberDestinationCache(): DestinationSearchResult | undefined {
     const destination = destinationCache.retrieveDestinationFromCache(
-      this.subscriberToken!.decoded,
+      this.selectSubscriberJwt(),
       this.name,
       this.options.isolationStrategy
     );
@@ -538,8 +537,8 @@ class DestinationFromServiceRetriever {
     return (
       this.subscriberToken &&
       isIdenticalTenant(
-        this.subscriberToken.decoded,
-        this.providerClientCredentialsToken.decoded
+        this.subscriberToken.serviceJwt.decoded,
+        this.providerServiceToken.decoded
       )
     );
   }