+++ test-packages/integration-tests/test/auth-flows/auth-flow.spec.ts
@@ -1,12 +1,9 @@
 import {
+  Destination,
   executeHttpRequest,
-  fetchDestination,
-  getDestination,
   getService,
-  serviceToken,
-  userApprovedServiceToken,
   wrapJwtInHeader,
-  jwtBearerToken,
+  getDestination,
   getDestinationFromDestinationService
 } from '@sap-cloud-sdk/core';
 import { BusinessPartner } from '@sap/cloud-sdk-vdm-business-partner-service';
@@ -17,7 +14,6 @@ import {
   Systems,
   UserAccessTokens
 } from './auth-flow-util';
-
 /* Consider the how-to-execute-auth-flow-tests.md to understand how to execute these tests. */
 
 describe('OAuth flows', () => {
@@ -32,55 +28,24 @@ describe('OAuth flows', () => {
     destinationService = getService('destination');
   });
 
-  xit('OAuth2SAMLBearerAssertion: Provider Destination & Provider Token', async () => {
-    const userGrant = await userApprovedServiceToken(
-      accessToken.provider,
-      destinationService
-    );
-
-    const destination = await fetchDestination(
-      destinationService!.credentials.uri,
-      userGrant,
-      systems.s4.providerOAuth2SAMLBearerAssertion
-    );
-    expect(destination.authTokens![0].error).toBeNull();
-
-    const result = await BusinessPartner.requestBuilder()
-      .getAll()
-      .top(1)
-      .execute(destination);
-    expect(result.length).toBe(1);
-  }, 60000);
-
   xit('OAuth2Password: Fetches destination and destination service has token', async () => {
-    const clientGrant = await serviceToken('destination', {
-      userJwt: accessToken.provider
-    });
-
-    const destination = await fetchDestination(
-      destinationService!.credentials.uri,
-      clientGrant,
-      systems.destination.providerOAuth2Password
+    const destination = await getDestination(
+      systems.destination.providerOauth2Password
     );
+
     expect(destination!.authTokens![0].type).toBe('bearer');
     expect(destination!.authTokens![0].value).toBeDefined();
   }, 60000);
 
   xit('BasicAuth: Provider Destination & Provider Token + GET request', async () => {
-    const clientGrant = await serviceToken('destination', {
+    const destination = await getDestination(systems.s4.providerBasic, {
       userJwt: accessToken.provider
     });
-
-    const destination = await fetchDestination(
-      destinationService!.credentials.uri,
-      clientGrant,
-      systems.s4.providerBasic
-    );
 
     const result = await BusinessPartner.requestBuilder()
       .getAll()
       .top(1)
-      .execute(destination);
+      .execute(destination!);
     expect(result.length).toBe(1);
   }, 60000);
 
@@ -99,15 +64,9 @@ describe('OAuth flows', () => {
   }, 60000);
 
   xit('BasicAuth: Provider Destination & Provider Token + PUT request (csrf token)', async () => {
-    const clientGrant = await serviceToken('destination', {
+    const destination = await getDestination(systems.s4.providerBasic, {
       userJwt: accessToken.provider
     });
-
-    const destination = await fetchDestination(
-      destinationService!.credentials.uri,
-      clientGrant,
-      systems.s4.providerBasic
-    );
 
     const buPa = BusinessPartner.builder()
       .businessPartnerCategory('1')
@@ -115,87 +74,234 @@ describe('OAuth flows', () => {
       .build();
     const result = await BusinessPartner.requestBuilder()
       .create(buPa)
-      .execute(destination);
+      .execute(destination!);
     expect(result.lastName).toBe('name');
   }, 60000);
 
   xit('BasicAuth: Subscriber Destination & Subscriber Token', async () => {
-    const clientGrant = await serviceToken('destination', {
+    const destination = await getDestination(systems.s4.subscriberBasic, {
       userJwt: accessToken.subscriber
     });
-
-    const destination = await fetchDestination(
-      destinationService!.credentials.uri,
-      clientGrant,
-      systems.s4.subscriberBasic
-    );
 
     const result = await BusinessPartner.requestBuilder()
       .getAll()
       .top(1)
-      .execute(destination);
+      .execute(destination!);
     expect(result.length).toBe(1);
   }, 60000);
 
-  xit('OAuth2ClientCredentials: Provider Destination & Provider Jwt', async () => {
-    const clientGrant = await serviceToken('destination', {
-      userJwt: accessToken.provider
+  xit('Basic Auth: iss as token ', async () => {
+    const destination = await getDestination(systems.s4.providerBasic, {
+      iss: 'http://s4sdk.localhost:8080/uaa/oauth/token'
     });
-
-    const destination = await fetchDestination(
-      destinationService!.credentials.uri,
-      clientGrant,
-      systems.workflow.providerOAuth2ClientCredentials
-    );
-    expect(destination.authTokens![0].error).toBeNull();
-
-    destination.url = destination.url + '/v1/workflow-definitions';
-    const response = await executeHttpRequest(destination, { method: 'get' });
-
-    expect(response.status).toBe(200);
+    expect(destination?.password).toBeDefined();
   }, 60000);
 
-  xit('OAuth2UserTokenExchange: Provider destination and Subscriber Jwt', async () => {
-    const providerDestToken = await serviceToken('destination', {});
-
-    const destination = await fetchDestination(
-      destinationService!.credentials.uri,
+  xit('OAuth2SAMLBearerAssertion: Provider Destination & Provider Token', async () => {
+    const destination = await getDestination(
+      systems.s4.providerOAuth2SAMLBearerAssertion,
       {
-        authHeaderJwt: providerDestToken,
-        exchangeHeaderJwt: accessToken.subscriber
-      },
-      systems.workflow.providerOAuth2UserTokenExchange
+        userJwt: accessToken.provider
+      }
     );
-
     expect(destination!.authTokens![0].error).toBeNull();
 
-    destination.url = destination.url + '/v1/workflow-definitions';
-    const response = await executeHttpRequest(destination, { method: 'get' });
+    const result = await BusinessPartner.requestBuilder()
+      .getAll()
+      .top(1)
+      .execute(destination!);
+    expect(result.length).toBe(1);
+  }, 60000);
+
+  xit('OAuth2ClientCredentials: Provider Destination (common token url)', async () => {
+    let destination = await getDestination(
+      systems.destination.providerOauth2ClientCredentialsCommonTokenURL,
+      {}
+    );
+    expect(destination!.authTokens![0]!.error).toBeNull();
+    destination = await getDestination(
+      systems.destination.providerOauth2ClientCredentialsCommonTokenURL,
+      {
+        userJwt: accessToken.subscriber
+      }
+    );
+    expect(destination!.authTokens![0]!.error).toBeNull();
+    destination = await getDestination(
+      systems.destination.providerOauth2ClientCredentialsCommonTokenURL,
+      {
+        userJwt: accessToken.provider
+      }
+    );
+    expect(destination!.authTokens![0]!.error).toBeNull();
+    assertCommenTokenUrl(destination!);
+  }, 60000);
+
+  xit('OAuth2ClientCredentials: Provider Destination (dedicated token service url)', async () => {
+    let destination = await getDestination(
+      systems.destination.providerOauth2ClientCredentials,
+      {}
+    );
+    expect(destination!.authTokens![0]!.error).toBeNull();
+    destination = await getDestination(
+      systems.destination.providerOauth2ClientCredentials,
+      {
+        userJwt: accessToken.provider
+      }
+    );
+    expect(destination!.authTokens![0]!.error).toBeNull();
+    destination = await getDestination(
+      systems.destination.providerOauth2ClientCredentials,
+      {
+        userJwt: accessToken.subscriber
+      }
+    );
+    expect(destination!.authTokens![0]!.error).toBeNull();
+    assertDedicatedTokenUrl(destination!);
+  }, 60000);
+
+  xit('OAuth2ClientCredentials: Provider Destination & Provider Jwt (workflow)', async () => {
+    const destination = await getDestination(
+      systems.workflow.providerOAuth2ClientCredentials,
+      {
+        userJwt: accessToken.provider
+      }
+    );
+    expect(destination!.authTokens![0].error).toBeNull();
+
+    destination!.url = destination!.url + '/v1/workflow-definitions';
+    const response = await executeHttpRequest(destination!, { method: 'get' });
 
     expect(response.status).toBe(200);
   }, 60000);
 
-  xit('OAuth2JWTBearer: Provider Destination & Provider Token', async () => {
-    const token = await serviceToken('destination', {
-      userJwt: accessToken.provider
-    });
-
-    const destination = await fetchDestination(
-      destinationService!.credentials.uri,
-      token,
-      systems.workflow.providerOauth2JWTBearer
+  xit('OAuth2UserTokenExchange: Provider destination and Provider Jwt (workflow)', async () => {
+    const destination = await getDestination(
+      systems.workflow.providerOAuth2UserTokenExchange,
+      {
+        userJwt: accessToken.provider
+      }
     );
 
     expect(destination!.authTokens![0].error).toBeNull();
 
-    destination.url = destination.url + '/v1/workflow-definitions';
-    const response = await executeHttpRequest(destination, { method: 'get' });
+    destination!.url = destination!.url + '/v1/workflow-definitions';
+    const response = await executeHttpRequest(destination!, { method: 'get' });
 
     expect(response.status).toBe(200);
+  }, 60000);
+
+  xit('OAuth2UserTokenExchange: Provider destination (common token url)', async () => {
+    let destination = await getDestination(
+      systems.destination.providerOauth2UserTokenExchangeCommonTokenURL,
+      {
+        userJwt: accessToken.subscriber
+      }
+    );
+    expect(destination!.authTokens![0].error).toBeNull();
+    destination = await getDestination(
+      systems.destination.providerOauth2UserTokenExchangeCommonTokenURL,
+      {
+        userJwt: accessToken.provider
+      }
+    );
+    expect(destination!.authTokens![0].error).toBeNull();
+    assertCommenTokenUrl(destination!);
+  }, 60000);
+
+  xit('OAuth2UserTokenExchange: Subscriber destination and Subscriber Jwt', async () => {
+    const destination = await getDestination(
+      systems.destination.subscriberOauth2UserTokenExchange,
+      {
+        userJwt: accessToken.subscriber
+      }
+    );
+    expect(destination!.authTokens![0].error).toBeNull();
+
+    const response = await executeHttpRequest(
+      {
+        url: 'https://destination-configuration.cfapps.sap.hana.ondemand.com/destination-configuration/v1/subaccountDestinations'
+      },
+      {
+        method: 'get',
+        headers: wrapJwtInHeader(destination!.authTokens![0].value).headers
+      }
+    );
+
+    expect(response.status).toBe(200);
+  }, 60000);
+
+  xit('OAuth2UserTokenExchange: Provider destination (dedicated token url)', async () => {
+    let destination = await getDestination(
+      systems.destination.providerOauth2UserTokenExchange,
+      {
+        userJwt: accessToken.provider
+      }
+    );
+    expect(destination!.authTokens![0].error).toBeNull();
+    destination = await getDestination(
+      systems.destination.providerOauth2UserTokenExchange,
+      {
+        userJwt: accessToken.subscriber
+      }
+    );
+    expect(destination!.authTokens![0].error).toMatch(
+      /Invalid issuer.*token did not match expected/
+    );
+    assertDedicatedTokenUrl(destination!);
+  }, 60000);
+
+  xit('OAuth2JWTBearer: Provider Destination & Provider Token (workflow call)', async () => {
+    const destination = await getDestination(
+      systems.workflow.providerOauth2JWTBearer,
+      {
+        userJwt: accessToken.provider
+      }
+    );
+
+    expect(destination!.authTokens![0].error).toBeNull();
+
+    destination!.url = destination!.url + '/v1/workflow-definitions';
+    const response = await executeHttpRequest(destination!, { method: 'get' });
+
+    expect(response.status).toBe(200);
+  }, 60000);
+
+  xit('OAuth2JWTBearer: Provider Destination  (common token service URL)', async () => {
+    let destination = await getDestination(
+      systems.destination.providerOauth2JWTBearerCommonTokenURL,
+      {
+        userJwt: accessToken.provider
+      }
+    );
+    expect(destination!.authTokens![0]!.error).toBeNull();
+    destination = await getDestination(
+      systems.destination.providerOauth2JWTBearerCommonTokenURL,
+      {
+        userJwt: accessToken.subscriber
+      }
+    );
+    expect(destination!.authTokens![0]!.error).toBeNull();
+  }, 60000);
+
+  xit('OAuth2JWTBearer: Provider Destination (dedicated token service URL)', async () => {
+    let destination = await getDestination(
+      systems.destination.providerOauth2JWTBearer,
+      {
+        userJwt: accessToken.subscriber
+      }
+    );
+    expect(destination!.authTokens![0]!.error).toMatch(/Unable to map issuer/);
+    destination = await getDestination(
+      systems.destination.providerOauth2JWTBearer,
+      {
+        userJwt: accessToken.provider
+      }
+    );
+    expect(destination!.authTokens![0]!.error).toBeNull();
   }, 60000);
 
   xit('ClientCertificate: Fetches the certificate and uses it', async () => {
-    const destination = await getDestination('CC8-HTTP-CERT');
+    const destination = await getDestination(systems.s4.providerClientCert, {});
     expect(destination!.certificates!.length).toBe(1);
     const bps = await BusinessPartner.requestBuilder()
       .getAll()
@@ -207,64 +313,51 @@ describe('OAuth flows', () => {
   // The axios does currently not support that the client cert socket is forwarded via a proxy:https://github.com/TooTallNate/node-https-proxy-agent/pull/111
   // Hence as a workaround it is possible to disable the proxy if it is not mandatory to pass the request.
   xit('ClientCertificate: Fetches the certificate and uses it - ignoring broken proxy', async () => {
-    process.env.HTTPS_PROXY = 'http://someHost:1234';
+    process.env.HTTPS_PROXY = 'http://someHost:1234'; // we changed to xssec which does not consider the no_proxy when calling the token. We would need a proper local forward everything proxy
     process.env.NO_PROXY =
       'https://s4sdk.authentication.sap.hana.ondemand.com/oauth/token,https://my300470-api.s4hana.ondemand.com';
 
-    const destination = await getDestination('CC8-HTTP-CERT');
+    const destination = await getDestination(systems.s4.providerClientCert, {});
     expect(destination!.certificates!.length).toBe(1);
     const bps = await BusinessPartner.requestBuilder()
       .getAll()
       .top(5)
       .execute(destination!);
     expect(bps.length).toBeGreaterThan(0);
-  }, 10000);
+  }, 60000);
 
-  xit('OAuth2UserTokenExchange: Subscriber destination and Subscriber Jwt', async () => {
-    const subscriberDestToken = await serviceToken('destination', {
-      userJwt: accessToken.subscriber
-    });
-
-    const destination = await fetchDestination(
-      destinationService!.credentials.uri,
+  xit('PrincipalPropagation: Provider not E2E', async () => {
+    const destination = await getDestination(
+      systems.s4onPrem.providerPrincipalPropagation,
       {
-        authHeaderJwt: subscriberDestToken,
-        exchangeHeaderJwt: accessToken.subscriber
-      },
-      systems.destination.subscriberOAuth2UserTokenExchange
-    );
-    expect(destination.authTokens![0].error).toBeNull();
-
-    const response = await executeHttpRequest(
-      {
-        url: 'https://destination-configuration.cfapps.sap.hana.ondemand.com/destination-configuration/v1/subaccountDestinations'
-      },
-      {
-        method: 'get',
-        headers: wrapJwtInHeader(destination.authTokens![0].value).headers
+        userJwt: accessToken.provider
       }
     );
-
-    expect(response.status).toBe(200);
-  }, 60000);
-
-  xit('OAuth2SAMLBearerAssertion: Provider Destination & Provider Token', async () => {
-    const jwtToken = await jwtBearerToken(
-      accessToken.provider,
-      destinationService
-    );
-
-    const destination = await fetchDestination(
-      destinationService!.credentials.uri,
-      jwtToken,
-      systems.s4.providerOAuth2SAMLBearerAssertion
-    );
-    expect(destination.authTokens![0].error).toBeNull();
-
-    const result = await BusinessPartner.requestBuilder()
-      .getAll()
-      .top(1)
-      .execute(destination);
-    expect(result.length).toBe(1);
+    // Call to backend will not work because the proxy is not rechable, but you can check the set headers.
+    expect(destination?.proxyConfiguration).toBeDefined();
+    expect(
+      destination?.proxyConfiguration?.headers?.['Proxy-Authorization']
+    ).toBeDefined();
+    expect(
+      destination?.proxyConfiguration?.headers?.[
+        'SAP-Connectivity-Authentication'
+      ]
+    ).toBeDefined();
   }, 60000);
 });
+
+function assertCommenTokenUrl(destination: Destination) {
+  expect(
+    destination.originalProperties!.destinationConfiguration[
+      'tokenServiceURLType'
+    ]
+  ).toBe('Common');
+}
+
+function assertDedicatedTokenUrl(destination: Destination) {
+  expect(
+    destination.originalProperties!.destinationConfiguration[
+      'tokenServiceURLType'
+    ]
+  ).toBe('Dedicated');
+}